Universal References for Function Objects (#328)
* Cleans up interfaces to operations that take a functor/lambda by using
universal references. This:
* allows fewer copies of function objects to be
made while preparing a message to send
* allows fewer copies while unpacking the lambda prior to its execution
* unifies async_visit and local_visit to allow for the same types of
function objects

In addition, a "test" is added that prints during each of a custom functor's
constructors to easily see the number of times functors are created.
This required a way of turning off the checks for lambda compliance to
use functors with non-default constructors. This should not be used
outside of debugging.